fix: Update constructor and test assertions for consistency in type usage

+++ src/Endorsable.sol
@@ -54,9 +54,8 @@ contract Endorsable is Ownable {
     /**
      * @dev constructor to ensure proper ownership is set - and set pre-requested addresses, if required. This contract is generally intended to be used via inheritance. Considering using `Ownable2step.sol` extension for a more secure ownership model.
      */
-    constructor( address[] memory _initialRequests ) Ownable(msg.sender) {
-
-        for (uint i = 0; i < _initialRequests.length; i++) {
+    constructor(address[] memory _initialRequests) Ownable(msg.sender) {
+        for (uint256 i = 0; i < _initialRequests.length; i++) {
             endorsements[_initialRequests[i]] = State.REQUESTED;
         }
     }
